refactor: Split store package into one file per collection
